Factor the code
