+++ src/applications/widget-editor/src/components/table-view/component.js
@@ -1,8 +1,7 @@
 import React, { useMemo } from "react";
 import PropTypes from "prop-types";
 
-import { CategoryIcon, NumberIcon, DateIcon, UnknownIcon } from '@widget-editor/shared';
-import AGGREGATION_OPTIONS from "@widget-editor/shared/lib/constants/aggregations";
+import { showFullTable, getAggregationName, getRelevantColumns } from "./helpers";
 import {
   StyledTableBox,
   StyledTable,
@@ -10,25 +9,6 @@ import {
   StyledTh,
   StyledTd,
 } from "./style";
-
-const getTypeIcon = (type) => {
-  let Icon;
-  switch (type) {
-    case 'string':
-      Icon = CategoryIcon;
-      break;
-    case 'number':
-      Icon = NumberIcon;
-      break;
-    case 'date':
-      Icon = DateIcon;
-      break;
-    default:
-      Icon = UnknownIcon;
-  }
-
-  return Icon;
-}
 
 const TableView = ({
   widgetData,
@@ -39,57 +19,14 @@ const TableView = ({
   color,
   aggregateFunction
 }) => {
-  const showFullTable = useMemo(() => !category || !value, [value, category]);
-
-  const aggregation = useMemo(() => aggregateFunction
-    ? AGGREGATION_OPTIONS.find(o => o.value === aggregateFunction)?.label
-    : null,
-  [aggregateFunction]);
+  const fullTable = useMemo(() => showFullTable(category, value), [value, category]);
+  const aggregation = useMemo(() => getAggregationName(aggregateFunction), [aggregateFunction]);
 
   // These are the colunms we want to show in the table
-  const relevantColumns = useMemo(() => {
-    // If the user has not set the colunms required to build a chart, we want to show a full table
-    if (showFullTable) {
-      const columns = columnOptions.reduce((res, option) => ({
-        ...res,
-        [option.value]: {
-          name: option.label,
-          Icon: getTypeIcon(option.type),
-        },
-      }), {});
-
-      return columns;
-    }
-
-    const columns = {
-      x: category ? { ...category, Icon: getTypeIcon(category.type) } : null,
-      y: value ? { ...value, Icon: getTypeIcon(value.type), aggregation } : null,
-      color: color ? { ...color, Icon: getTypeIcon(color.type) } : null,
-    };
-
-    // We remove the columns that are not set by the user
-    Object.keys(columns).forEach(key => {
-      if (!columns[key]) {
-        delete columns[key];
-      }
-    });
-
-    // If the user has chosen the same column for the X and Y axis, and there is not an aggregation,
-    // we avoid displaying the column twice in the table
-    if (columns.x?.name && columns.x?.name === columns.y?.name && !aggregation) {
-      delete columns.y;
-    }
-
-    // If the user has chosen the same column for the X axis and color, or the user has chosen the
-    // same column for the Y axis and color and there is not an aggregation, we also remove the
-    // color column to avoid displaying it twice in the table
-    if ((columns.x?.name && columns.x?.name === columns.color?.name)
-      || (columns.y?.name && columns.y?.name === columns.color?.name && !aggregation)) {
-      delete columns.color;
-    }
-
-    return columns;
-  }, [category, value, color, aggregation, showFullTable, columnOptions]);
+  const relevantColumns = useMemo(
+    () => getRelevantColumns(category, value, color, aggregation, columnOptions),
+    [category, value, color, aggregation, columnOptions]
+  );
 
   return (
     <StyledTableBox>
@@ -113,7 +50,7 @@ const TableView = ({
           </StyledTr>
         </thead>
         <tbody>
-          {(showFullTable ? tableData : widgetData).map((row, index) => (
+          {(fullTable ? tableData : widgetData).map((row, index) => (
             <StyledTr key={index}>
               {Object.keys(relevantColumns).map(
                 column => <StyledTd key={column}>{row[column]}</StyledTd>
